win_requirements.txt
absl-py==2.4.0
accelerate==1.12.0
albucore==0.0.23
albumentations==1.4.24
alembic==1.18.4
annotated-types==0.7.0
certifi==2026.1.4
charset-normalizer==3.4.4
colorama==0.4.6
colorlog==6.10.1
contourpy==1.3.3
cycler==0.12.1
defusedxml==0.7.1
filelock==3.24.3
filetype==1.2.0
fonttools==4.61.1
fsspec==2026.2.0
greenlet==3.3.2
grpcio==1.78.1
huggingface_hub==0.36.2
idna==3.7
Jinja2==3.1.6
kiwisolver==1.4.9
Mako==1.3.10
Markdown==3.10.2
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.2
opencv-python==4.13.0.92
opencv-python-headless==4.10.0.84
optuna==4.7.0
packaging==26.0
pandas==3.0.1
peft==0.18.1
pi_heif==1.2.1
pillow==12.1.1
pillow-avif-plugin==1.5.5
protobuf==6.33.5
psutil==7.2.2
pycocotools==2.0.11
pydantic==2.12.5
pydantic_core==2.41.5
pyDeprecate==0.4.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2026.2.19
requests==2.32.5
requests-toolbelt==1.0.0
rf100vl==1.1.0
rfdetr==1.5.1
rfdetr-plus==1.0.1
roboflow==1.2.14
safetensors==0.7.0
scipy==1.17.0
seaborn==0.13.2
simsimd==6.5.13
six==1.17.0
SQLAlchemy==2.0.47
stringzilla==4.6.0
supervision==0.27.0.post1
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tokenizers==0.22.2
torch @ file:///C:/Users/AIBuser/Desktop/kd_to_delete_FINAL_Complete/Optuna_training-rf-detr/torch-2.8.0%2Bcu126-cp311-cp311-win_amd64.whl#sha256=cfb2c640a8955fbd8686c056802f53a512b610c098960d6dad5800cbc16c02b6
torchvision @ file:///C:/Users/AIBuser/Desktop/kd_to_delete_FINAL_Complete/Optuna_training-rf-detr/torchvision-0.23.0%2Bcu126-cp311-cp311-win_amd64.whl#sha256=921ec706184a4b79190483724d439fb2f769b4cde83e3031a7b7ac097676a63b
tqdm==4.67.3
transformers==4.57.6
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.1.6


